Updated parsed content
Now specify if content is `html` or `md` the current set up makes it so is something is in `html` (i.e. day templates and unit overviews) the inner tag content gets wrapped in a `<p>` tag. This update checks if a file is already in `html` meaning there is no apparent need to proccess