Enable locking modules with prerequsites
The overall support structure was already present, all that was needed
was a parser to implement the ability to put a completiton requirement
on a given item.

Then in order to allow other modules to require previous modules to be
completed first, all that was needed was a module item parser, and an
additional attribute supported on the module tag. Because we are using
id's for everything now this makes support for it much easier to
implement.

Relevant documentation is located
[here](https://developerdocs.instructure.com/services/canvas/resources/modules)

Signed-off-by: Ethan Dye <mrtops03@gmail.com>